requirements.txt
python-dotenv==1.0.0
watchdog==3.0.0
oracledb==1.3.0
pdfplumber==0.10.0
pdfminer.six==20221105
pdf2image==1.16.0
pillow==10.0.0
pytesseract==0.3.10
ultralytics==8.0.200
opencv-python==4.8.0.76
numpy==1.26.4
pywin32==306
openai
